Created basis for movie provider
